requirements.txt
altair
numpy
pandas
pydeck
streamlit
Office365-REST-Python-Client
openpyxl